requirements.txt
Flask==2.2.5
requests==2.32.0
gunicorn==23.0.0
coverage==7.6.9
python-dotenv==0.19.0
folium==0.19.2
python-dateutil==2.9.0
pytz==2024.2